requirements.txt
pan-os-python
toml